+++ dotfiles/anyos/nvim/init.lua
@@ -1,4 +1,2 @@
 -- bootstrap lazy.nvim, LazyVim and your plugins
 require("config.lazy")
--- plain native statusline (lualine is disabled in lua/plugins/lualine.lua)
-require("config.statusline")